pyproject.toml
[project]
name = "duck-thursday-protocol"
version = "0.1.0"
description = "Local Thursday assistant with Telegram and OpenCode bridge."
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
]

[project.scripts]
thursday = "thursday.app.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
